Split core image generation module
